+++ extern/SAASound/src/SAANoise.h
@@ -10,25 +10,25 @@
 class CSAANoise  
 {
 private:
-	unsigned long m_nCounter;
-	unsigned long m_nAdd;
-	unsigned long m_nCounter_low;
+	unsigned int m_nCounter;
+	uint64_t m_nAdd;
+	unsigned int m_nCounter_low;
 	unsigned int m_nOversample;
-	unsigned long m_nCounterLimit_low;
+	unsigned int m_nCounterLimit_low;
 	bool m_bSync; // see description of "SYNC" bit of register 28
-	unsigned long m_nSampleRate; // = 44100 when RateMode=0, for example
+	uint64_t m_nSampleRate; // = 44100 when RateMode=0, for example
 	int m_nSourceMode;
-	unsigned long m_nAddBase; // nAdd for 31.25 kHz noise at 44.1 kHz samplerate
+	uint64_t m_nAddBase; // nAdd for 31.25 kHz noise at 44.1 kHz samplerate
 
 	// pseudo-random number generator
-	unsigned long m_nRand;
+	unsigned int m_nRand;
 
 	void ChangeLevel(void);
 	
 
 public:
 	CSAANoise();
-	CSAANoise(unsigned long seed);
+	CSAANoise(unsigned int seed);
 	~CSAANoise();
 
 	void SetSource(int nSource);
@@ -36,7 +36,7 @@ public:
 	void _SetSampleRate(int nSampleRate);
 	void _SetOversample(unsigned int oversample);
 	void _SetClockRate(int nClockRate);
-	void Seed(unsigned long seed);
+	void Seed(unsigned int seed);
 
 	void Tick(void);
 	int Level(void) const;
